+++ .github/workflows/ci.yml
@@ -16,12 +16,17 @@ jobs:
       fail-fast: false
       matrix:
         spark: [
-            "3.0.0", "3.0.1", "3.0.2",
-            "3.1.0", "3.1.1", "3.1.2",
-            "3.2.0"
+          "2.4.0","2.4.1","2.4.2",
+          "2.4.3","2.4.4","2.4.5",
+          "2.4.6","2.4.7","2.4.8",
+          "3.0.0", "3.0.1", "3.0.2",
+          "3.1.0", "3.1.1", "3.1.2",
+          "3.2.0"
         ]
         scope: [ "test" ]
         include:
+          - spark: "2.4"
+            scope: "test-uploadReport"
           - spark: "3.0"
             scope: "test-uploadReport"
           - spark: "3.1"
