+++ .github/workflows/release.yml
@@ -2,8 +2,8 @@ name: Release
 
 on:
   push:
-    branches: [main]
-    tags: ["*"]
+    branches: [ main ]
+    tags: [ "*" ]
 
 concurrency:
   group: ${{ github.workflow }}-${{ github.ref }}
@@ -22,7 +22,7 @@ jobs:
     strategy:
       fail-fast: true
       matrix:
-        spark: [ "3.0", "3.1", "3.2" ]
+        spark: [ "2.4", "3.0", "3.1", "3.2" ]
     runs-on: ubuntu-20.04
     steps:
       - uses: actions/checkout@v3
