+++ .github/workflows/wc_publishSite.yml
@@ -13,7 +13,7 @@ jobs:
     strategy:
       fail-fast: false
       matrix:
-        spark: [ "3.0", "3.1", "3.2" ]
+        spark: [ "2.4", "3.0", "3.1", "3.2" ]
         scope: [ "${{ inputs.scope }}" ]
     runs-on: ubuntu-20.04
     steps:
