+++ core/src/main/docs/scala/doric/doc/DDataframe.scala
@@ -1,8 +1,6 @@
 package doric
 package doc
 
-import org.apache.spark.sql.DoricDataframe
-
 /**
   * Dummy class to show all the extension methods for the dataframe created in
   * Doric documentation This class cant be created, extended and should not be
