+++ core/src/main/scala/doric/syntax/CommonColumns.scala
@@ -4,8 +4,9 @@ package syntax
 import cats.implicits._
 import doric.sem.Location
 import doric.types.{Casting, SparkType, UnsafeCasting}
-import org.apache.spark.sql.catalyst.expressions.ArrayRepeat
+
 import org.apache.spark.sql.{Column, functions => f}
+import org.apache.spark.sql.catalyst.expressions.ArrayRepeat
 
 private[syntax] trait CommonColumns extends ColGetters[NamedDoricColumn] {
 
@@ -34,16 +35,6 @@ private[syntax] trait CommonColumns extends ColGetters[NamedDoricColumn] {
     */
   def hash(cols: DoricColumn[_]*): IntegerColumn =
     cols.map(_.elem).toList.sequence.map(f.hash(_: _*)).toDC
-
-  /**
-    * Calculates the hash code of given columns using the 64-bit
-    * variant of the xxHash algorithm, and returns the result as a long column.
-    *
-    * @group All Types
-    * @see [[org.apache.spark.sql.functions.xxhash64]]
-    */
-  def xxhash64(cols: DoricColumn[_]*): LongColumn =
-    cols.map(_.elem).toList.sequence.map(f.xxhash64(_: _*)).toDC
 
   /**
     * Returns the least value of the list of values, skipping null values.
