+++ core/src/main/spark_2.4_mount/scala/doric/syntax/All.scala
@@ -1,0 +1,21 @@
+package doric.syntax
+
+private[doric] trait All
+    extends ArrayColumns
+    with TypeMatcher
+    with CommonColumns
+    with DStructs
+    with LiteralConversions
+    with MapColumns
+    with NumericColumns
+    with DateColumns
+    with TimestampColumns
+    with BooleanColumns
+    with StringColumns
+    with ControlStructures
+    with AggregationColumns
+    with CNameOps
+    with BinaryColumns
+    with Interpolators
+    with BinaryColumns30_31
+    with StringColumn24
