+++ core/src/main/spark_3.1_mount/scala/doric/syntax/All.scala
@@ -23,3 +23,7 @@ private[doric] trait All
     with NumericColumns31
     with StringColumns31
     with BinaryColumns30_31
+    with ArrayColumns3x
+    with CommonColumns3x
+    with MapColumns3x
+    with StringColumn3x
