+++ core/src/main/spark_3.2_mount/scala/doric/syntax/All.scala
@@ -22,3 +22,7 @@ private[doric] trait All
     with NumericColumns31
     with StringColumns31
     with BinaryColumns32
+    with ArrayColumns3x
+    with CommonColumns3x
+    with MapColumns3x
+    with StringColumn3x
